requirements.txt
Django==1.6.5
jingo==0.7
South==1.0
MySQL-python==1.2.5
feedparser==5.1.3
pytz
bleach==1.4
readability-lxml=0.3.0.3
